requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
aioredis==1.2.0
aiosmtplib==1.1.4
amqp==5.0.2
asn1crypto==1.4.0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
base58==2.1.0
bidict==0.21.2
billiard==3.6.3.0
bitcoin==1.1.42
bitcoinaddress==0.1.5
bitcoinlib==0.5.2
bitsv==0.11.5
CacheControl==0.12.6
cachetools==4.1.1
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coincurve==15.0.0
colorama==0.4.4
cryptography==3.2.1
cryptos==1.36
dicttoxml==1.7.4
ecdsa==0.16.0
firebase-admin==4.4.0
flatten-dict==0.2.0
getmac==0.8.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h11==0.9.0
hiredis==1.1.0
httpcore==0.11.1
httplib2==0.18.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
isort==5.6.4
Jinja2==2.11.2
jwt==1.1.0
kombu==5.0.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
motor==2.3.0
msgpack==1.0.0
multidict==5.0.0
Naked==0.1.31
numpy==1.19.5
pathlib2==2.3.5
Pillow==8.0.1
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.14.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pyfcm==1.4.7
PyJWT==1.7.1
pylint==2.6.0
-e git://github.com/ojii/pymaging.git@596a08fce5664e58d6e8c96847393fbe987783f2#egg=pymaging
-e git://github.com/ojii/pymaging-png.git@83d85c44e4b2342818e6c068065e031a9f81bb9f#egg=pymaging_png
pymongo==3.11.1
python-bitcoinrpc==1.0
python-bitcointx==1.1.1.post2
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
redis==3.5.3
requests==2.25.0
rfc3986==1.4.0
rsa==4.6
sanic==20.9.1
sanic-base-extension==0.1.1
sanic-compress==0.1.1
Sanic-Cors==0.10.0.post3
sanic-jinja2==0.8.0
sanic-jwt-extended==1.0.dev12
sanic-openapi==0.6.2
Sanic-Plugins-Framework==0.9.4.post1
sanic-redis==0.2.0
sanic-redis-extension==0.1.1
sanic-scheduler==1.0.7
scrypt==0.8.17
shellescape==3.8.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
toml==0.10.2
twilio==6.48.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
uheprng==1.3
ujson==4.0.1
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
whatsonchain==0.0.3
wrapt==1.12.1
yarl==1.6.3